code climate fixes
